pyproject.toml
[tool.poetry]
name = "python-polylith-test"
version = "0.1.0"
description = ""
authors = ["Christian Packard <6462409+cpackard@users.noreply.github.com>"]
readme = "README.md"
packages = [
         {include = "development"},
         {include = "cpackard/choices", from = "components"},
         {include = "cpackard/questions", from = "components"},
         {include = "cpackard/authentication", from = "components"},
         {include = "cpackard/api", from = "bases"}
]
[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2022.7.1"
django-extensions = {version = "^3.2.1"}
pytest = "^7.2.2"
pytest-django = "^4.5.2"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cpackard.api.settings"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
import_heading_stdlib = "Standard Library"
import_heading_thirdparty = "Third-Party Libraries"
import_heading_django = "Django Libraries"
import_heading_firstparty = "Polylith Bricks"
import_heading_localfolder = "Local Modules"
known_firstparty = [ "cpackard",]
known_django = [ "django", "rest_framework",]
sections = [ "FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER",]

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
follow_imports = "skip"

[tool.mypy.plugins.django-stubs]
django_settings_module = "cpackard.api.settings"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.0.1"
isort = "^5.12.0"
requests = "^2.28.2"
django-stubs = "^1.16.0"
django-types = "^0.17.0"
flake8 = "^6.0.0"

[tool.poetry.group.emacs.dependencies]
python-lsp-server = "^1.7.1"
python-lsp-black = "^1.2.1"
pylsp-mypy = "^0.6.6"
autoflake = "^2.0.1"
importmagic = "^0.1.7"
epc = "^0.0.5"
werkzeug = "^2.2.3"
debugpy = "^1.6.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
